requirements.txt
streamlit
anthropic
python-dotenv
psycopg2-binary
